Reduser and actions seprated from index.js to modules dir.
