+++ public/index.html
@@ -15,6 +15,7 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div class="root jumbotron">
+      <h1 class="title">Counter:</h1>
       <h2 class="counter">0</h2>
       <button class="dec btn btn-primary btn-lg">
         DEC
